send/requirements.txt
pyzmq==19.0.1
pycrypto==2.6.1
json2xml==3.4.1